refactor: modularize character manager
